requirements.txt
Pillow==10.0.0
PyQt6==6.5.2